Week_07/Twitter/slackbot/requirements.txt
slackclient==2.8.0
SQLAlchemy
psycopg2-binary
